refactor chembl lookup
