+++ mrtarget/common/chembl_lookup.py
@@ -11,14 +11,6 @@ from elasticsearch_dsl.query import Match
 from opentargets_urlzsource import URLZSource
 import simplejson as json
 
-
-def get_chembl_info_by_file(uri):
-    with URLZSource(uri).open() as f_obj:
-        for i, line in enumerate(f_obj, start=1):
-            cheml_dict = json.loads(line)
-            yield cheml_dict
-
-
 class ChEMBLLookup(object):
     def __init__(self, target_uri, mechanism_uri, component_uri, protein_uri, 
             molecule_set_uri_pattern):
@@ -45,9 +37,8 @@ class ChEMBLLookup(object):
         self.protein_class_label_to_id = {}
         self.molecules_dict = self.populate_molecules_dict()
 
-
     def populate_molecules_dict(self):
-        self._logger.info('ChEMBL getting Molecule from ' + self.molecule_set_uri_pattern)
+        self._logger.debug('ChEMBL getting Molecule from ' + self.molecule_set_uri_pattern)
         # Shelve creates a file with specific database. Using a temp file requires a workaround to open it.
         # dumbdbm creates an empty database file. In this way shelve can open it properly.
         t_filename = tempfile.NamedTemporaryFile(delete=False).name
@@ -58,51 +49,38 @@ class ChEMBLLookup(object):
                 mol = json.loads(line)
                 shelve_out[str(mol["molecule_chembl_id"])] = mol
 
-        self._logger.info('ChEMBL Molecule loading done. ')
+        self._logger.debug('ChEMBL Molecule loading done. ')
         return shelve_out
 
-
-    def download_targets(self):
-        '''fetches all the targets from chembl and store their data and a mapping to uniprot id'''
-
-        self._logger.info('ChEMBL getting targets from ' +
-                          self.target_uri)
-
-        targets = get_chembl_info_by_file(self.target_uri)
-        for i in targets:
-            if 'target_components' in i and \
-                    i['target_components'] and \
-                            'accession' in i['target_components'][0] and \
-                    i['target_components'][0]['accession']:
-                uniprot_id = i['target_components'][0]['accession']
-                self.targets[uniprot_id] = i
-                self.uni2chembl[uniprot_id] = i['target_chembl_id']
-
-    def download_mechanisms(self):
-        '''fetches mechanism data and stores which molecules are
-        linked to any target'''
-
-        mechanisms = get_chembl_info_by_file(self.mechanism_uri)
-
-        allowed_target_chembl_ids = set(self.uni2chembl.values())
-        for i in mechanisms:
-            self.mechanisms[i['record_id']] = i
-            target_id = i['target_chembl_id']
-            if target_id in allowed_target_chembl_ids:
-                if target_id not in self.target2molecule:
-                    self.target2molecule[target_id] = set()
-                self.target2molecule[target_id].add(i['molecule_chembl_id'])
 
     def download_molecules_linked_to_target(self):
         '''generate a dictionary with all the synonyms known for a given molecules.
          Only retrieves molecules linked to a target'''
-        self._logger.info('chembl downloading molecules linked to target')
-        if not self.targets:
-            self._logger.debug('chembl downloading targets')
-            self.download_targets()
-        if not self.target2molecule:
-            self._logger.debug('chembl downloading mechanisms')
-            self.download_mechanisms()
+
+        '''fetches all the targets from chembl and store their data and a mapping to uniprot id'''
+
+        with URLZSource(self.target_uri).open() as f_obj:
+            for line in f_obj:
+                i = json.loads(line)
+                if 'target_components' in i and \
+                        i['target_components'] and \
+                        'accession' in i['target_components'][0] and \
+                        i['target_components'][0]['accession']:
+                    uniprot_id = i['target_components'][0]['accession']
+                    self.targets[uniprot_id] = i
+                    self.uni2chembl[uniprot_id] = i['target_chembl_id']
+
+        allowed_target_chembl_ids = set(self.uni2chembl.values())
+        with URLZSource(self.mechanism_uri).open() as f_obj:
+            for line in f_obj:
+                i = json.loads(line)
+                self.mechanisms[i['record_id']] = i
+                target_id = i['target_chembl_id']
+                if target_id in allowed_target_chembl_ids:
+                    if target_id not in self.target2molecule:
+                        self.target2molecule[target_id] = set()
+                    self.target2molecule[target_id].add(i['molecule_chembl_id'])
+
         required_molecules = set()
         self._logger.info('chembl t2m mols')
         for molecules in self.target2molecule.values():
@@ -117,28 +95,22 @@ class ChEMBLLookup(object):
 
     def download_protein_classification(self):
         '''fetches targets components from chembls and inject the target class data in self.protein_classification'''
-        self.download_protein_class()
 
-        targets_components = get_chembl_info_by_file(self.component_uri)
-        for i in targets_components:
-            if 'accession' in i:
-                if i['accession'] not in self.protein_classification:
-                    self.protein_classification[i['accession']] = []
-                for classification in i['protein_classifications']:
-                    protein_class_id = classification['protein_classification_id']
-                    self.protein_classification[i['accession']].append(self.protein_class[protein_class_id])
+        with URLZSource(self.protein_uri).open() as f_obj:
+            for line in f_obj:
+                i = json.loads(line)
+                protein_class_id = i.pop('protein_class_id')
+                protein_class_data = dict((k, dict(label=v, id='')) for k, v in i.items() if v)  # remove values with none
+                self.protein_class[protein_class_id] = protein_class_data
 
-    def download_protein_class(self):
-        '''
-        Fetches target classes from chembl and stores it in self.protein_class
-        :return:
-        '''
-        protein_classes = get_chembl_info_by_file(self.protein_uri)
-        for i in protein_classes:
-            protein_class_id = i.pop('protein_class_id')
-            protein_class_data = dict((k, dict(label=v, id='')) for k, v in i.items() if v)  # remove values with none
-            self.protein_class[protein_class_id] = protein_class_data
-            self._store_label_to_protein_class(protein_class_id, protein_class_data)
+                max_level = 0
+                label = ''
+                for k, v in protein_class_data.items():
+                    level = int(k[1])
+                    if level >= max_level:
+                        max_level = level
+                        label = v['label']
+                self.protein_class_label_to_id[label] = protein_class_id
 
         '''inject missing ids'''
         for k, v in self.protein_class.items():
@@ -147,15 +119,15 @@ class ChEMBLLookup(object):
                 if label in self.protein_class_label_to_id:
                     data['id'] = self.protein_class_label_to_id[label]
 
-    def _store_label_to_protein_class(self, protein_class_id, protein_class_data):
-        max_level = 0
-        label = ''
-        for k, v in protein_class_data.items():
-            level = int(k[1])
-            if level >= max_level:
-                max_level = level
-                label = v['label']
-        self.protein_class_label_to_id[label] = protein_class_id
+        with URLZSource(self.component_uri).open() as f_obj:
+            for line in f_obj:
+                i = json.loads(line)
+                if 'accession' in i:
+                    if i['accession'] not in self.protein_classification:
+                        self.protein_classification[i['accession']] = []
+                    for classification in i['protein_classifications']:
+                        protein_class_id = classification['protein_classification_id']
+                        self.protein_classification[i['accession']].append(self.protein_class[protein_class_id])
 
     def get_molecules_from_evidence(self, es, index):
 
@@ -203,7 +175,6 @@ class ChEMBLLookup(object):
             if self.molecules_dict.has_key(mol_k):
                 data['molecules'].append(self.molecules_dict[mol_k])
             else:
-                self._logger.error('problem retrieving the molecule info from the local db %s', str(mol_k))
                 raise ValueError('problem retrieving the molecule info from the local db', str(mol_k))
 
         #if the data is what we expected, process it
